Separated client file, batch return requests, cancel messages
